粗鄙不堪！

+++ src/java/iot/service/OrderService.java
@@ -15,6 +15,7 @@ import iot.dao.repository.CustomerPriceDAO;
 import iot.dao.repository.OrderDetailDAO;
 import iot.dao.repository.OrderHeadDAO;
 import iot.dao.repository.ProductDAO;
+import iot.dao.repository.exceptions.NonexistentEntityException;
 import iot.dao.repository.exceptions.PreexistingEntityException;
 import javax.persistence.EntityManagerFactory;
 import iot.response.Response;
@@ -116,6 +117,7 @@ public class OrderService {
                     list_row.add(od.getProductMasterId().getProductName());
                     list_row.add(od.getOrderQty());
                     list_row.add(od.getOrderPrice());
+                    list_row.add(od.getVersionNumber());
                     responseList.add(list_row);
                 }
             }
@@ -223,12 +225,17 @@ public class OrderService {
      * 刪除訂單頭檔
      *
      * @param orderHeadId
+     * @param versionNumber
      * @return
      * @throws PreexistingEntityException
      */
-    public Response deleteOrderHeadService(String orderHeadId) throws PreexistingEntityException {
+    public Response deleteOrderHeadService(String orderHeadId,String versionNumber) throws PreexistingEntityException, NonexistentEntityException {
         OrderHeadDAO ohdao = new OrderHeadDAO(emf);
         OrderHead oh = ohdao.queryOrderHeadByOrderHeadId(orderHeadId);
+        if (oh==null) {
+            throw new NonexistentEntityException("編號爲:"+oh.getOrderHeadId()+" 的訂單頭檔已不存在！");
+        }
+        oh.setVersionNumber(Integer.parseInt(versionNumber));
         return ohdao.deleteOrderHead(oh);
     }
 
@@ -236,12 +243,19 @@ public class OrderService {
      * 刪除訂單身檔
      *
      * @param orderDetailId
+     * @param versionNumber
      * @return
      * @throws PreexistingEntityException
+     * @throws iot.dao.repository.exceptions.NonexistentEntityException
      */
-    public Response deleteOrderDetailService(String orderDetailId) throws PreexistingEntityException {
+    public Response deleteOrderDetailService(String orderDetailId,String versionNumber) throws PreexistingEntityException, NonexistentEntityException {
         OrderDetailDAO oddao = new OrderDetailDAO(emf);
         OrderDetail od = (OrderDetail) oddao.queryOrderDetailByOrderDetailId(orderDetailId).getData();
+         if (od==null) {
+            throw new NonexistentEntityException("編號爲:"+od.getOrderDetailId()+" 的訂單身檔已不存在！");
+        }
+        od.setVersionNumber(Integer.parseInt(versionNumber));
+        
         return oddao.deleteOrderDetail(od);
     }
